Refactor(#19) : User 관련 코드 구조 정리 및 필요없는 개행 삭제
