+++ src/main/java/com/keodam/keodam_backend/app/user/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.keodam.keodam_backend.app.user.controller;
 
-import com.keodam.keodam_backend.app.user.domain.User;
+import com.keodam.keodam_backend.app.domain.User;
 import com.keodam.keodam_backend.app.user.dto.NicknameRequestDto;
 import com.keodam.keodam_backend.app.user.dto.RoleRequestDto;
 import com.keodam.keodam_backend.app.user.dto.SignupStatusResponseDto;
