+++ src/main/java/com/keodam/keodam_backend/app/user/repository/UserRepository.java
@@ -1,10 +1,8 @@
 package com.keodam.keodam_backend.app.user.repository;
 
-import com.keodam.keodam_backend.app.user.domain.User;
-
+import com.keodam.keodam_backend.app.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
