+++ src/main/java/com/keodam/keodam_backend/app/user/service/UserService.java
@@ -1,7 +1,7 @@
 package com.keodam.keodam_backend.app.user.service;
 
 import com.keodam.keodam_backend.app.domain.RoleType;
-import com.keodam.keodam_backend.app.user.domain.User;
+import com.keodam.keodam_backend.app.domain.User;
 import com.keodam.keodam_backend.app.user.dto.SignupStatusResponseDto;
 import com.keodam.keodam_backend.app.user.repository.UserRepository;
 import com.keodam.keodam_backend.app.user.dto.UserResponseDto;
@@ -10,7 +10,6 @@ import com.keodam.keodam_backend.global.code.status.ErrorStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 import java.util.Optional;
