+++ src/main/java/com/keodam/keodam_backend/global/code/BaseErrorCode.java
@@ -3,5 +3,4 @@ package com.keodam.keodam_backend.global.code;
 public interface BaseErrorCode {
     ErrorReasonDTO getReason();
     ErrorReasonDTO getReasonHttpStatus();
-
 }
