+++ src/main/java/com/keodam/keodam_backend/global/code/status/ErrorStatus.java
@@ -56,11 +56,7 @@ public enum ErrorStatus implements BaseErrorCode {
     NICKNAME_CONTAINS_BAD_WORD(HttpStatus.BAD_REQUEST, "NICKNAME4005", "닉네임에 비속어는 포함할 수 없어요."),
 
     // role type
-    INVALID_ROLE_TYPE(HttpStatus.BAD_REQUEST, "ROLE4001", "잘못된 역할 선택입니다."),
-
-
-    //quiz
-    QUIZ_NOT_FOUND(HttpStatus.NOT_FOUND,"QUIZ4001","퀴즈를 찾을 수 없습니다.");
+    INVALID_ROLE_TYPE(HttpStatus.BAD_REQUEST, "ROLE4001", "잘못된 역할 선택입니다.");
 
     private final HttpStatus httpStatus;
     private final String code;
