+++ src/main/java/com/keodam/keodam_backend/global/security/JwtAuthenticationProcessingFilter.java
@@ -1,6 +1,6 @@
 package com.keodam.keodam_backend.global.security;
 
-import com.keodam.keodam_backend.app.user.domain.User;
+import com.keodam.keodam_backend.app.domain.User;
 import com.keodam.keodam_backend.app.user.repository.UserRepository;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
