+++ src/main/java/com/keodam/keodam_backend/mypage/service/FileService.java
@@ -1,6 +1,6 @@
 package com.keodam.keodam_backend.mypage.service;
 
-import com.keodam.keodam_backend.app.user.domain.User;
+import com.keodam.keodam_backend.app.domain.User;
 import com.keodam.keodam_backend.app.user.repository.UserRepository;
 import com.keodam.keodam_backend.exception.GeneralException;
 import com.keodam.keodam_backend.global.aws.AwsS3Service;
