+++ src/main/java/com/keodam/keodam_backend/oauth/controller/AuthTestController.java
@@ -2,9 +2,7 @@ package com.keodam.keodam_backend.oauth.controller;
 
 import com.keodam.keodam_backend.app.domain.RoleType;
 import com.keodam.keodam_backend.app.domain.SocialType;
-
-
-import com.keodam.keodam_backend.app.user.domain.User;
+import com.keodam.keodam_backend.app.domain.User;
 import com.keodam.keodam_backend.app.user.repository.UserRepository;
 import com.keodam.keodam_backend.global.security.JwtService;
 import io.swagger.v3.oas.annotations.Operation;
