+++ src/main/java/com/keodam/keodam_backend/oauth/domain/IdTokenAttributes.java
@@ -2,7 +2,7 @@ package com.keodam.keodam_backend.oauth.domain;
 
 import com.keodam.keodam_backend.app.domain.RoleType;
 import com.keodam.keodam_backend.app.domain.SocialType;
-import com.keodam.keodam_backend.app.user.domain.User;
+import com.keodam.keodam_backend.app.domain.User;
 import com.keodam.keodam_backend.oauth.domain.userinfo.AppleUserInfo;
 import com.keodam.keodam_backend.oauth.domain.userinfo.GoogleUserInfo;
 import com.keodam.keodam_backend.oauth.domain.userinfo.KakaoUserInfo;
