+++ src/main/java/com/keodam/keodam_backend/oauth/service/IdTokenService.java
@@ -3,7 +3,7 @@ package com.keodam.keodam_backend.oauth.service;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.keodam.keodam_backend.app.domain.SocialType;
-import com.keodam.keodam_backend.app.user.domain.User;
+import com.keodam.keodam_backend.app.domain.User;
 import com.keodam.keodam_backend.app.user.repository.UserRepository;
 import com.keodam.keodam_backend.oauth.domain.CustomIdTokenUser;
 import com.keodam.keodam_backend.oauth.domain.IdTokenAttributes;
